refactor(bus-core): move Module management in bus-core module
#10
