+++ packages/evtbus/src/index.ts
@@ -1,5 +1,4 @@
 import { Worker } from 'worker_threads'
-import * as bus  from '@bsorrentino/bus-core'
 import assert = require('assert')
 import {
     Broker, 
@@ -14,53 +13,6 @@ import {
  */
 const broker = new Broker()
 
-
-/**
- * Module information
- */
- export type ModuleInfo = { module:bus.Module, status:bus.ModuleStatus }
-
-/**
- * Module Management
- */
-export class Modules {
-
-    private _modules = new Map<string,ModuleInfo>()
-
-    register<C extends bus.ModuleConfiguration>( module:bus.Module<C>, config?:C  ) {
-        assert.ok( !this._modules.has( module.name ), `Module ${module.name} already exists!` )
-
-        let result:ModuleInfo = {
-            module:module,
-            status:{ started:false, paused:false} 
-        }
-        this._modules.set( module.name, result )
-        if( module.onRegister ) {
-            module.onRegister( config )
-        }
-    }
-    
-    get names():IterableIterator<string> {
-        return this._modules.keys()
-    }
-
-    start() {
-        this._modules.forEach( m => {
-
-            if( !m.status.started ) {
-                if( m.module.onStart ) {
-                    m.module.onStart()
-                }
-                m.status.started = true
-            }
-        })
-    }
-}
-
-/**
- * global modules instance
- */
-export const modules = new Modules()
 
 /**
  * get all names of the instantiated channels
